requirements.txt
amqp==5.0.6
attrs==21.2.0
billiard==3.6.4.0
cachetools==4.2.2
celery==5.1.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
environ-config==21.2.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Pydantic==0.8.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.31.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
kubernetes==17.17.0
MarkupSafe==2.0.1
oauthlib==3.1.1
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
smmap==4.0.0
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.1.0
Werkzeug==2.0.1
